# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Healthy Home Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /canvassing/sessions:
    get:
      operationId: listCanvassingSessions
      tags: [canvassing]
      summary: List canvassing sessions
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: canvasser
          in: query
          schema:
            type: string
        - name: neighborhood
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of canvassing sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CanvassingSession"
    post:
      operationId: createCanvassingSession
      tags: [canvassing]
      summary: Create a canvassing session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCanvassingSession"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanvassingSession"
  /canvassing/sessions/{id}:
    get:
      operationId: getCanvassingSession
      tags: [canvassing]
      summary: Get a canvassing session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Canvassing session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanvassingSession"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCanvassingSession
      tags: [canvassing]
      summary: Update a canvassing session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCanvassingSession"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanvassingSession"
  /canvassing/leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, quoted, follow_up, sold, lost]
        - name: canvasser
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLead"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /canvassing/leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLead"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomer"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomer"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    CanvassingSession:
      type: object
      properties:
        id:
          type: integer
        canvasser:
          type: string
        sessionDate:
          type: string
        neighborhood:
          type: string
          nullable: true
        route:
          type: string
          nullable: true
        doorsKnocked:
          type: integer
        peopleReached:
          type: integer
        goodConversations:
          type: integer
        quotesGiven:
          type: integer
        closes:
          type: integer
        revenueSold:
          type: string
        averageTicket:
          type: string
          nullable: true
        bundleCount:
          type: integer
        driveawayAddOnCount:
          type: integer
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - canvasser
        - sessionDate
        - doorsKnocked
        - peopleReached
        - goodConversations
        - quotesGiven
        - closes
        - revenueSold
        - bundleCount
        - driveawayAddOnCount
        - createdAt
        - updatedAt
    CreateCanvassingSession:
      type: object
      properties:
        canvasser:
          type: string
        sessionDate:
          type: string
        neighborhood:
          type: string
          nullable: true
        route:
          type: string
          nullable: true
        doorsKnocked:
          type: integer
        peopleReached:
          type: integer
        goodConversations:
          type: integer
        quotesGiven:
          type: integer
        closes:
          type: integer
        revenueSold:
          type: string
        averageTicket:
          type: string
          nullable: true
        bundleCount:
          type: integer
        driveawayAddOnCount:
          type: integer
        notes:
          type: string
          nullable: true
      required:
        - canvasser
        - sessionDate
        - doorsKnocked
        - peopleReached
        - goodConversations
        - quotesGiven
        - closes
        - revenueSold
        - bundleCount
        - driveawayAddOnCount
    Lead:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        address:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        leadSource:
          type: string
          nullable: true
        canvasser:
          type: string
          nullable: true
        quoteAmount:
          type: string
          nullable: true
        serviceInterest:
          type: string
          nullable: true
        status:
          type: string
          enum: [new, quoted, follow_up, sold, lost]
        followUpDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        sessionId:
          type: integer
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - customerName
        - status
        - createdAt
        - updatedAt
    CreateLead:
      type: object
      properties:
        customerName:
          type: string
        address:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        leadSource:
          type: string
          nullable: true
        canvasser:
          type: string
          nullable: true
        quoteAmount:
          type: string
          nullable: true
        serviceInterest:
          type: string
          nullable: true
        status:
          type: string
          enum: [new, quoted, follow_up, sold, lost]
        followUpDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        sessionId:
          type: integer
          nullable: true
      required:
        - customerName
    Customer:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zip:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        optOut:
          type: boolean
        reviewCampaignEligible:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - firstName
        - lastName
        - optOut
        - reviewCampaignEligible
        - createdAt
        - updatedAt
    CreateCustomer:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zip:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        optOut:
          type: boolean
        reviewCampaignEligible:
          type: boolean
      required:
        - firstName
        - lastName
...