# lib/api-zod/src/generated/types/createCanvassingSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Healthy Home Operating System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCanvassingSession {
  canvasser: string;
  sessionDate: string;
  neighborhood?: string | null;
  route?: string | null;
  doorsKnocked: number;
  peopleReached: number;
  goodConversations: number;
  quotesGiven: number;
  closes: number;
  revenueSold: string;
  averageTicket?: string | null;
  bundleCount: number;
  driveawayAddOnCount: number;
  notes?: string | null;
}